backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
google-generativeai>=0.3.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.28.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pytest>=7.0.0
psycopg2-binary
